requirements.txt
aiomysql==0.2.0
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
databases==0.8.0
fastapi==0.108.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
MarkupSafe==2.1.3
mysql-connector-python==8.2.0
protobuf==4.21.12
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.51
SQLAlchemy-Utils==0.41.1
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
Werkzeug==3.0.1
